# openapi/panel-api.yaml
openapi: 3.0.3
info:
  title: 3x-ui Panel API
  version: "1.0.0"
  description: |
    OpenAPI specification for the bearer-authenticated panel API exposed under `/panel/api`.

    Notes:
    - All documented endpoints can be called with `Authorization: Bearer <token>`.
    - The implementation also accepts `X-API-Key: <token>` as an alternative.
    - If the panel is served behind a custom `webBasePath`, prepend that base path before `/panel/api`.
    - Most mutation endpoints return HTTP 200 even when `success=false` inside the JSON envelope.
servers:
  - url: "{scheme}://{host}{basePath}"
    description: Panel origin plus optional configured web base path.
    variables:
      scheme:
        default: http
        enum:
          - http
          - https
      host:
        default: localhost:2053
      basePath:
        default: ""
        description: Custom panel base path without trailing slash, for example `/my-panel`.
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /panel/api/inbounds/list:
    get:
      tags: [Inbounds]
      summary: List inbounds visible to the authenticated user
      operationId: listInbounds
      responses:
        "200":
          description: Inbound list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          $ref: "#/components/schemas/Inbound"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/get/{id}:
    get:
      tags: [Inbounds]
      summary: Get one inbound by ID
      operationId: getInbound
      parameters:
        - $ref: "#/components/parameters/InboundId"
      responses:
        "200":
          description: Inbound record.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/Inbound"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/getClientTraffics/{email}:
    get:
      tags: [Inbounds]
      summary: Get client traffic by email
      operationId: getClientTrafficByEmail
      parameters:
        - $ref: "#/components/parameters/Email"
      responses:
        "200":
          description: Client traffic record or null.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        nullable: true
                        allOf:
                          - $ref: "#/components/schemas/ClientTraffic"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/getClientTrafficsById/{id}:
    get:
      tags: [Inbounds]
      summary: Get all client traffic rows for an inbound
      operationId: getClientTrafficByInboundId
      parameters:
        - $ref: "#/components/parameters/InboundIdAsString"
      responses:
        "200":
          description: Client traffic rows for the specified inbound.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          $ref: "#/components/schemas/ClientTraffic"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/vkTurnProxy/peerOptions:
    get:
      tags: [Inbounds]
      summary: List WireGuard peers available for binding to a vk-turn-proxy inbound
      operationId: getVkTurnProxyPeerOptions
      parameters:
        - $ref: "#/components/parameters/InboundId"
      responses:
        "200":
          description: Peer options for the linked WireGuard inbound.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/VKTurnProxyPeerOptionsResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/vkTurnProxy/clients:
    post:
      tags: [Inbounds]
      summary: Create one vk-turn-proxy client
      operationId: createVkTurnProxyClient
      parameters:
        - $ref: "#/components/parameters/InboundId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VKTurnProxyClient"
      responses:
        "200":
          description: Client created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/vkTurnProxy/clients/{clientId}/stats:
    get:
      tags: [Inbounds]
      summary: Get traffic stats for one vk-turn-proxy client
      operationId: getVkTurnProxyClientStats
      parameters:
        - $ref: "#/components/parameters/InboundId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: Traffic stats for the selected client.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        nullable: true
                        allOf:
                          - $ref: "#/components/schemas/ClientTraffic"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/vkTurnProxy/clients/{clientId}/enable:
    post:
      tags: [Inbounds]
      summary: Set vk-turn-proxy client enabled state
      operationId: setVkTurnProxyClientEnable
      parameters:
        - $ref: "#/components/parameters/InboundId"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VKTurnProxyClientEnableRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/VKTurnProxyClientEnableRequest"
      responses:
        "200":
          description: Updated enabled state.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/EnableStateResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/vkTurnProxy/exportClient/{clientId}:
    get:
      tags: [Inbounds]
      summary: Export one vk-turn-proxy client as a wingsv link
      operationId: exportVkTurnProxyClient
      parameters:
        - $ref: "#/components/parameters/InboundId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: "One `wingsv://` import link."
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: string
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/vkTurnProxy/exportAll:
    get:
      tags: [Inbounds]
      summary: Export all vk-turn-proxy clients as wingsv links
      operationId: exportAllVkTurnProxyClients
      parameters:
        - $ref: "#/components/parameters/InboundId"
      responses:
        "200":
          description: Exported client links.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: array
                        items:
                          $ref: "#/components/schemas/VKTurnProxyExportedClient"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/add:
    post:
      tags: [Inbounds]
      summary: Create an inbound
      operationId: addInbound
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundWriteRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/InboundWriteRequest"
      responses:
        "200":
          description: Creation result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/Inbound"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/del/{id}:
    post:
      tags: [Inbounds]
      summary: Delete an inbound
      operationId: deleteInbound
      parameters:
        - $ref: "#/components/parameters/InboundId"
      responses:
        "200":
          description: Deletion result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        type: integer
                        format: int64
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/update/{id}:
    post:
      tags: [Inbounds]
      summary: Update an inbound
      operationId: updateInbound
      parameters:
        - $ref: "#/components/parameters/InboundId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundWriteRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/InboundWriteRequest"
      responses:
        "200":
          description: Update result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        $ref: "#/components/schemas/Inbound"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/clientIps/{email}:
    post:
      tags: [Inbounds]
      summary: Get stored client IP addresses
      operationId: getClientIps
      parameters:
        - $ref: "#/components/parameters/Email"
      responses:
        "200":
          description: IP record list or string.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      obj:
                        oneOf:
                          - type: string
                          - type: array
                            items:
                              type: string
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/clearClientIps/{email}:
    post:
      tags: [Inbounds]
      summary: Clear stored client IP addresses
      operationId: clearClientIps
      parameters:
        - $ref: "#/components/parameters/Email"
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/addClient:
    post:
      tags: [Inbounds]
      summary: Add client records to an inbound
      operationId: addInboundClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientMutationRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ClientMutationRequest"
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/addPeer:
    post:
      tags: [Inbounds]
      summary: Add WireGuard peers to an inbound
      operationId: addInboundPeer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeerMutationRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PeerMutationRequest"
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/delClient/{clientId}:
    post:
      tags: [Inbounds]
      summary: Delete a client by client identifier
      description: |
        `clientId` is protocol-specific:
        - VMess/VLESS: client UUID
        - Trojan: password
        - Shadowsocks: email
        URL-encode the path value if it contains special characters.
      operationId: deleteInboundClient
      parameters:
        - $ref: "#/components/parameters/InboundId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
  /panel/api/inbounds/{id}/delPeer/{peerId}:
    post:
      tags: [Inbounds]
      summary: Delete a WireGuard peer by index
      operationId: deleteInboundPeer
      parameters:
        - $ref: "#/components/parameters/InboundId"
        - $ref: "#/components/parameters/PeerIndex"
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
...
  /panel/api/inbounds/{id}/resetClientTraffic/{email}:
    post:
      tags: [Inbounds]
      summary: Reset one client's traffic counters
      operationId: resetClientTraffic
      parameters:
        - $ref: "#/components/parameters/InboundId"
        - $ref: "#/components/parameters/Email"
      responses:
        "200":
          $ref: "#/components/responses/ApiEnvelopeResponse"
        "404":
          description: Missing or invalid bearer token.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
      description: "Pass the panel API token as `Authorization: Bearer <token>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative header accepted by the implementation.
  parameters:
    InboundId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Email:
      name: email
      in: path
      required: true
      schema:
        type: string
      description: Client email. URL-encode the value when needed.
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        msg:
          type: string
        obj:
          nullable: true
...